refactor: reorganize components into feature-based directories
- Create new directories: diary/, mood/, notifications/, a11y/
- Move 13 components from root to feature-specific directories:
  - diary/: DiaryDeleteDialog, DiaryDetailModal, DiaryFilter, DiaryPreview
  - mood/: MoodSelector, EmotionTagChips, EmotionTagManager, EmotionTagSelector
  - notifications/: GlobalNotification, NotificationSettings, ReminderDialog
  - a11y/: SkipLink
  - pagination/: PaginationComponent
- Update all import paths in src/ and tests/
- No code changes, only file reorganization

Closes #261
