+++ src/components/mood/EmotionTagChips.vue
