+++ src/components/mood/EmotionTagManager.vue
