+++ src/components/mood/EmotionTagSelector.vue
