+++ src/components/mood/MoodSelector.vue
