+++ src/components/pagination/PaginationComponent.vue
