+++ src/views/DiaryViewPage.vue
@@ -145,7 +145,8 @@ import { useDataStore } from '@/stores/data'
 import { useAuthStore } from '@/stores/auth'
 import { useEmotionTagsStore } from '@/stores/emotionTags'
 import { useDiaries } from '@/composables/useDataFetch'
-import EmotionTagChips from '@/components/EmotionTagChips.vue'
+import DiaryFilter from '@/components/diary/DiaryFilter.vue'
+import EmotionTagChips from '@/components/mood/EmotionTagChips.vue'
 import type { DiaryEntry, DiaryEntryWithEmotionTags } from '@/types/custom'
 import { createLogger } from '@/utils/logger'
