+++ tests/components/normal_EmotionTagSelector_01.spec.ts
@@ -7,7 +7,7 @@ import { createPinia, setActivePinia } from 'pinia'
 import { createVuetify } from 'vuetify'
 import * as components from 'vuetify/components'
 import * as directives from 'vuetify/directives'
-import EmotionTagSelector from '@/components/EmotionTagSelector.vue'
+import EmotionTagSelector from '@/components/mood/EmotionTagSelector.vue'
 import { useEmotionTagsStore } from '@/stores/emotionTags'
 import type { EmotionTag } from '@/types/emotion-tags'
